fix small bug in cache filename generation
git-svn-id: http://subversion.ffado.org/ffado/branches/libffado-2.0@1244 2be59082-3212-0410-8809-b0798e1608f0

+++ src/bebob/bebob_avdevice.h
@@ -105,7 +105,7 @@ public:
 protected:
     virtual uint8_t getConfigurationIdSampleRate();
     virtual uint8_t getConfigurationIdNumberOfChannel( AVC::PlugAddress::EPlugDirection ePlugDirection );
-    virtual uint8_t getConfigurationIdSyncMode();
+    virtual uint16_t getConfigurationIdSyncMode();
 
 protected:
     Mixer*             m_Mixer;
